refactor(step5): 패키지 분리
