+++ src/main/kotlin/racingcar/domain/car/AMGEngine.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 import kotlin.random.Random
