+++ src/main/kotlin/racingcar/domain/car/CarFactory.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 object CarFactory {
     fun getCars(carNames: List<String>): List<Car> = carNames.map { Car(it) }
