+++ src/main/kotlin/racingcar/domain/car/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 interface Engine {
     val rpm: Int
