+++ src/main/kotlin/racingcar/ui/input/InputValidator.kt
@@ -1,4 +1,4 @@
-package racingcar.input
+package racingcar.ui.input
 
 class InputValidator {
     fun isNumber(value: String) = value.matches("\\d+".toRegex())
