+++ src/test/kotlin/racingcar/domain/car/CarTest.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
