+++ src/test/kotlin/racingcar/domain/car/MockEngine.kt
@@ -1,3 +1,3 @@
-package racingcar.car
+package racingcar.domain.car
 
 class MockEngine(override val rpm: Int) : Engine
