+++ src/test/kotlin/racingcar/ui/input/InputValidatorTest.kt
@@ -1,4 +1,4 @@
-package racingcar.input
+package racingcar.ui.input
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
